dedicated arithmetic mixin
